python-generators-0x00/requirements.txt
mysql-connector==2.2.9
python-decouple==3.8
